backend/requirements.txt
asgiref==3.8.1
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==4.2.18
django-extra-views==0.14.0
django-haystack==3.3.0
django-import-export==4.3.4
django-oscar==3.2.5
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory-boy==3.2.1
Faker==35.0.0
idna==3.10
packaging==24.2
phonenumbers==8.13.53
pillow==11.1.0
purl==1.6
pycountry==24.6.1
pysolr==3.10.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sorl-thumbnail==12.9.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
